refactor: split Python core.py into focused modules
Split the monolithic core.py (1284 lines) into 5 smaller modules:
- _ffi.py: FFI loading, C constants, CDEF
- sasa.py: SASA calculation (single + batch)
- classifier.py: atom classification and radius lookup
- rsa.py: RSA calculation
- batch.py: directory batch processing

Public API unchanged via __init__.py re-exports.
