+++ python/zsasa/xtc.py
@@ -34,7 +34,8 @@ from typing import TYPE_CHECKING, Literal
 import numpy as np
 from numpy.typing import NDArray
 
-from zsasa.core import _get_lib, calculate_sasa_batch
+from zsasa._ffi import _get_lib
+from zsasa.sasa import calculate_sasa_batch
 
 if TYPE_CHECKING:
     from cffi import FFI
