+++ website/docs/guide/classifiers.mdx
@@ -92,7 +92,7 @@ result = calculate_sasa(coords, classification.radii)
   <TabItem value="python-gemmi" label="Python (Gemmi)">
 
 ```python
-from zsasa.core import ClassifierType
+from zsasa import ClassifierType
 from zsasa.integrations.gemmi import calculate_sasa_from_structure
 
 # Classifier is applied automatically
